8.14
BugFix

+++ src/controllers/videoController.js
@@ -20,16 +20,28 @@ export const watch = async (req, res) => {
 };
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not Found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { pageTitle: `Editing: ${video.title}`, video });
 };
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.exists({ _id: id });
   const { title, description, hashtags } = req.body;
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   if (!video) {
     return res.restatus(404).direct(`404`, { pageTitle: "Video not found." });
   }
@@ -75,6 +87,18 @@ export const postUpload = async (req, res) => {
 };
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", {
+      pageTitle: "Video not Found",
+    });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).render("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
